lang-importer/project/requirements.txt
streamlit==1.31.0
openai==4.28.0
python-dotenv==1.0.0